[Policies] extracts PoliciesForm from PolicyConfig
