+++ spec/javascripts/Policies/components/PolicyConfig.spec.jsx
@@ -1,7 +1,8 @@
 import React from 'react'
 import Enzyme, { mount } from 'enzyme'
 import Adapter from 'enzyme-adapter-react-16'
-import { PolicyConfig, PolicyForm } from 'Policies/components/PolicyConfig'
+import { PolicyConfig } from 'Policies/components/PolicyConfig'
+import { PolicyForm } from 'Policies/components/PoliciesForm'
 
 Enzyme.configure({ adapter: new Adapter() })
